Feat/vbd amortize (#409)
* raw untested changes

* cleanup artifact

* tensor conversion fix

* working amortization script

* amortization fixes

* replace examples with vbd

* fixes

* revert to 32 agents

* variable agent count

* amortized womd agents
