+++ src/init.hpp
@@ -37,6 +37,7 @@ namespace madrona_gpudrive
         uint32_t id;
         MapVector2 mean;
         bool markAsExpert{false};
+        float vbd_trajectories[consts::episodeLen][5];  // x, y, yaw, vx, vy
     };
 
     struct MapRoad
