+++ src/mgr.hpp
@@ -73,6 +73,7 @@ public:
     MGR_EXPORT madrona::py::Tensor deletedAgentsTensor() const;
     MGR_EXPORT madrona::py::Tensor mapNameTensor() const;
     MGR_EXPORT madrona::py::Tensor scenarioIdTensor() const;
+    MGR_EXPORT madrona::py::Tensor vbdTrajectoryTensor() const;
     madrona::py::Tensor rgbTensor() const;
     madrona::py::Tensor depthTensor() const;
     // These functions are used by the viewer to control the simulation
