+++ src/sim.cpp
@@ -59,6 +59,7 @@ void Sim::registerTypes(ECSRegistry &registry, const Config &cfg)
     registry.registerComponent<RoadMapId>();
     registry.registerComponent<MapType>();
     registry.registerComponent<MetaData>();
+    registry.registerComponent<VBDTrajectory>();
 
     registry.registerSingleton<WorldReset>();
     registry.registerSingleton<Shape>();
@@ -117,6 +118,8 @@ void Sim::registerTypes(ECSRegistry &registry, const Config &cfg)
         (uint32_t)ExportID::Trajectory);
     registry.exportColumn<AgentInterface, MetaData>(
         (uint32_t)ExportID::MetaData);
+    registry.exportColumn<AgentInterface, VBDTrajectory>(
+        (uint32_t)ExportID::VBDTrajectory);
 }
 
 static inline void cleanupWorld(Engine &ctx) {
