+++ src/sim.hpp
@@ -41,6 +41,7 @@ enum class ExportID : uint32_t {
     DeletedAgents,
     MapName,
     ScenarioId,
+    VBDTrajectory,
     NumExports
 };
