+++ src/types.hpp
@@ -362,6 +362,28 @@ namespace madrona_gpudrive
 
     static_assert(sizeof(Trajectory) == sizeof(float) * TrajectoryExportSize);
 
+    struct VBDTrajectory
+    {
+        // For each agent, store the full VBD trajectory (x, y, yaw, vx, vy)
+        // The tensor has shape [traj_length, 5]
+        float trajectories[consts::episodeLen][5];
+
+        static inline void zero(VBDTrajectory& vbd_traj)
+        {
+            for (int i = 0; i < consts::episodeLen; i++)
+            {
+                for (int j = 0; j < 5; j++)
+                {
+                    vbd_traj.trajectories[i][j] = 0.0f;
+                }
+            }
+        }
+    };
+
+    const size_t VBDTrajectoryExportSize = consts::episodeLen * 5;
+
+    static_assert(sizeof(VBDTrajectory) == sizeof(float) * VBDTrajectoryExportSize);
+
     struct Shape
     {
         int32_t agentEntityCount;
@@ -449,6 +471,7 @@ namespace madrona_gpudrive
                                 Trajectory,
                                 AgentID,
                                 MetaData,
+                                VBDTrajectory,
 
                                 ControlledState // Drive Logic
